refactor: change package name to com.easylaw.app
- Update applicationId in app/build.gradle
- Update namespace in app/build.gradle
- Refactor all Kotlin files to use new package name
- Refactor all test files to use new package name
- Delete old com.neobrix.easylaw package folders
- Update imports in all source files

Package name: com.neobrix.easylaw → com.easylaw.app
