+++ app/src/main/java/com/easylaw/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.neobrix.easylaw
+package com.easylaw.app
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.neobrix.easylaw.ui.theme.EasyLawTheme
+import com.easylaw.app.ui.theme.EasyLawTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 /**
