+++ app/src/main/java/com/easylaw/app/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.neobrix.easylaw.ui.theme
+package com.easylaw.app.ui.theme
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.lightColorScheme
